// web/src/api/generated/models/catalogManagementMenuResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * POS Cafe API
 * Backend API for Cafe Point of Sale System built with Vertical Slice Architecture and Idiomatic Go.
 * OpenAPI spec version: 1.0
 */
import type { CatalogManagementCategoryResponse } from './catalogManagementCategoryResponse';

export interface CatalogManagementMenuResponse {
  categories?: CatalogManagementCategoryResponse[];
}